+++ flutter_inappwebview_windows/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.7.0-beta.3
 
 - Updated flutter_inappwebview_platform_interface version to ^1.4.0-beta.3
+- Implemented `preferredColorScheme` property of `InAppWebViewSettings` to control WebView2 color scheme (Light/Dark/Auto) via `ICoreWebView2Profile.PreferredColorScheme` (requires WebView2 Runtime 1.0.1901.177+)
 - Merged "windows: fix WebViewEnvironment dispose crash" [#2433](https://github.com/pichillilorenzo/flutter_inappwebview/pull/2433) (thanks to [GooRingX](https://github.com/GooRingX))
 - Merged "fix #2484, Remove not-empty assert for Cookie.value" [#2486](https://github.com/pichillilorenzo/flutter_inappwebview/pull/2486) (thanks to [laishere](https://github.com/laishere))
 - Merged "Prevent Unpredictable Close On Windows" [#2543](https://github.com/pichillilorenzo/flutter_inappwebview/pull/2543) (thanks to [momadvisor](https://github.com/momadvisor))